requirements.txt
openai
python-dotenv
ollama
flask
markdown 